+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_vserver_peer_permissions.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2023-2025, NetApp, Inc
+# (c) 2023-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -39,6 +39,30 @@ options:
       - Specifies name of the peer Cluster.
     type: str
     required: true
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality. See the README file for examples.
 """
 
 EXAMPLES = """
@@ -97,9 +121,12 @@ class NetAppONTAPVserverPeerPermissions:
             cluster_peer=dict(required=True, type='str')
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[['use_lambda', True, ('lambda_config',)]],
         )
 
         # set up variables