+++ plugins/modules/na_ontap_name_mappings.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2022-2025, NetApp, Inc
+# (c) 2022-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
@@ -84,6 +84,30 @@ options:
       - Minimum value is 1 and maximum is 2147483647.
       - Requires ONTAP version 9.7 or later.
     type: int
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality. See the README file for examples.
 
 '''
 
@@ -158,9 +182,14 @@ class NetAppOntapNameMappings:
             replacement=dict(required=False, type='str'),
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
         self.na_helper = NetAppModule(self)
         self.parameters = self.na_helper.set_parameters(self.module.params)
@@ -252,6 +281,7 @@ class NetAppOntapNameMappings:
         reindex = False
         current = self.get_name_mappings_rest()
         cd_action = self.na_helper.get_cd_action(current, self.parameters)
+
         # Throws error when trying to swap with non existing index
         if cd_action == 'create':
             if self.parameters.get('from_index') is not None: