+++ plugins/modules/na_ontap_name_service_switch.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2019-2025, NetApp, Inc
+# (c) 2019-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -9,6 +9,7 @@ __metaclass__ = type
 
 DOCUMENTATION = '''
 author: NetApp Ansible Team (@carchi8py) <ng-ansible-team@netapp.com>
+short_description: "NetApp ONTAP Manage name service switch"
 description:
   - Create/Delete/Modify Name Service Switch.
   - Deleting name service switch not supported in REST.
@@ -39,8 +40,30 @@ options:
       - Possible values include files,dns,ldap,nis.
     type: list
     elements: str
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 
-short_description: "NetApp ONTAP Manage name service switch"
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -91,9 +114,14 @@ class NetAppONTAPNsswitch:
             sources=dict(required=False, type='list', elements='str')
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -106,6 +134,8 @@ class NetAppONTAPNsswitch:
         self.use_rest = self.rest_api.is_rest()
         self.svm_uuid = None
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])